Alpine training: use MariaDB instead of MySQL.
The mysql image has changed to be oracle based, so commands used in
the training such as `ps` or `top` are not existing in the image. The
training cannot be completed.

The `debian` version might work, but does not contain a multi arch
build so it does not work on ARM based processors, such as newer
MacBooks with M1/M2 chips.

The best solution is to simply use the MariaDB opensource version,
that was forked from MySQL some time ago.